requirements.txt
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.9
amqp==5.1.1
asgiref==3.7.1
Babel==2.12.1
billiard==3.6.4.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
dbfread==2.0.7
Django==4.2.1
django-extensions==3.2.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
isodate==0.6.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
leather==0.3.4
MarkupSafe==2.1.2
numpy==1.24.3
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parsedatetime==2.6
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
gunicorn==20.1.0
django_celery_beat
redis==4.5.5
celery[redis]==5.2.7
django-cors-headers==4.0.0
